+++ lib/demoflu.dart
@@ -1,8 +1,7 @@
 library demoflu;
 
 export 'src/demoflu_app.dart';
-export 'src/example.dart';
-export 'src/section.dart';
+export 'src/menu_item.dart';
 export 'src/menu/example_menu_notifier.dart';
 export 'src/menu/example_menu_widgets.dart';
 export 'src/menu/example_stateful.dart';