+++ lib/src/menu/app_menu_widget.dart
@@ -1,6 +1,5 @@
 import 'package:demoflu/src/demoflu_app.dart';
-import 'package:demoflu/src/example.dart';
-import 'package:demoflu/src/section.dart';
+import 'package:demoflu/src/menu_item.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
@@ -10,35 +9,24 @@ class AppMenuWidget extends StatelessWidget {
   Widget build(BuildContext context) {
     DemoFluAppState state = DemoFluAppState.of(context)!;
     List<Widget> children = [];
-    for (Section section in state.sections) {
-      if (section.name != null) {
-        children.add(Container(
-            child: Text(section.name!,
-                style: TextStyle(
-                    fontStyle: FontStyle.italic, color: Colors.grey[300])),
-            padding: EdgeInsets.only(left: 8, right: 16, top: 8)));
-      }
-      for (Example example in section.examples) {
-        children.add(_MenuItem(state.currentExample == example, example));
-      }
+    for (MenuItem menuItem in state.menuItems) {
+      children.add(_MenuItem(state.currentMenuItem == menuItem, menuItem));
     }
     return Container(
         child: SingleChildScrollView(
-            child: Column(
-                children: children,
-                crossAxisAlignment: CrossAxisAlignment.start)),
-        decoration: BoxDecoration(
-            color: Colors.blueGrey[800],
-            border: Border(
-                right: BorderSide(width: 1, color: Colors.blueGrey[900]!))));
+            child: IntrinsicWidth(
+                child: Column(
+                    children: children,
+                    crossAxisAlignment: CrossAxisAlignment.stretch))),
+        decoration: BoxDecoration(color: Colors.blueGrey[900]));
   }
 }
 
 class _MenuItem extends StatefulWidget {
-  const _MenuItem(this.selected, this.example);
+  const _MenuItem(this.selected, this.menuItem);
 
   final bool selected;
-  final Example example;
+  final MenuItem menuItem;
 
   @override
   State<StatefulWidget> createState() => _MenuItemState();
@@ -49,22 +37,45 @@ class _MenuItemState extends State<_MenuItem> {
 
   @override
   Widget build(BuildContext context) {
+    double left = 8;
+    if (widget.menuItem.indentation > 1) {
+      left = left * widget.menuItem.indentation;
+    }
+    Widget text = Padding(
+        child: Text(widget.menuItem.name,
+            style: TextStyle(color: _textColor(), fontStyle: _fontStyle())),
+        padding: EdgeInsets.fromLTRB(left, 8, 8, 8));
+
+    if (widget.menuItem.example == null) {
+      return text;
+    }
+
     return MouseRegion(
         onEnter: _onEnter,
         onExit: _onExit,
         child: Container(
-            margin: EdgeInsets.only(top: 8, bottom: 0, right: 16),
-            decoration: BoxDecoration(
-                border:
-                    Border(left: BorderSide(width: 8, color: _borderColor()))),
-            child: _text()));
+            color: _background(), child: InkWell(child: text, onTap: _onTap)));
   }
 
-  Color _borderColor() {
+  FontStyle _fontStyle() {
+    if (widget.menuItem.italic) {
+      return FontStyle.italic;
+    }
+    return FontStyle.normal;
+  }
+
+  Color _textColor() {
     if (widget.selected) {
-      return Colors.white;
+      return Colors.black;
+    }
+    return Colors.white;
+  }
+
+  Color _background() {
+    if (widget.selected) {
+      return Colors.blueGrey[100]!;
     } else if (hover) {
-      return Colors.white.withOpacity(.4);
+      return Colors.blueGrey[700]!;
     }
     return Colors.transparent;
   }
@@ -81,17 +92,8 @@ class _MenuItemState extends State<_MenuItem> {
     });
   }
 
-  Widget _text() {
-    return InkWell(
-        child: Padding(
-            child: Text(widget.example.name,
-                style: TextStyle(color: Colors.white)),
-            padding: EdgeInsets.only(left: 8)),
-        onTap: _onTap);
-  }
-
   _onTap() {
     DemoFluAppState state = DemoFluAppState.of(context)!;
-    state.updateCurrentExample(widget.example);
+    state.updateCurrentExample(widget.menuItem);
   }
 }